Refactor queues
