+++ backend/src/queues/index.ts
@@ -1,0 +1,3 @@
+import GeorgeQueue from "./george-queue";
+
+export default [GeorgeQueue];
