+++ backend/src/server.ts
@@ -11,7 +11,7 @@ import { Server } from "http";
 import { version } from "./version";
 import { recordServerVersion } from "./utils/prometheus";
 import * as RedisClient from "./init/redis";
-import { initJobQueue } from "./tasks/george";
+import queues from "./queues";
 import Logger from "./utils/logger";
 
 async function bootServer(port: number): Promise<Server> {
@@ -41,7 +41,9 @@ async function bootServer(port: number): Promise<Server> {
       Logger.success("Connected to redis");
 
       Logger.info("Initializing task queues...");
-      initJobQueue(RedisClient.getConnection());
+      queues.forEach((queue) => {
+        queue.init(RedisClient.getConnection());
+      });
       Logger.success("Task queues initialized");
     }
